requirements.txt
audioread==2.1.4 
 
decorator==4.0.10  
 
Flask==0.11.1  
 
future==0.16.0  
   
 
Jinja2==2.8 
 
joblib==0.10.3  
 
librosa==0.5.0rc0  
 
matplotlib==1.4.3  
 
numpy==1.11.3  
 
Pillow==3.4.2 
requests==2.12.4  
 
scikit-learn==0.18.1  
 
scipy==0.18.1  
 
six==1.10.0  
 
virtualenv==15.1.0
Werkzeug==0.11.13 
win-unicode-console==0.5 
 
